backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
Flask-JWT-Extended
mysql-connector-python
python-dotenv
email-validator
